Rework bottom bar to use a new IconButton component that renders our sag icons correctly. Move icons for bottom bar into a dedicated folder Move Wind Direction and Speed controls into a placeholder terrain-panel component (to be completed) Currently, interaction with the spark button renders the lower buttons inactive, with no way to reset until spark placement is completed.
