+++ src/components/icon-button.scss
@@ -1,0 +1,34 @@
+@import "common.scss";
+
+.iconButton{
+  min-width: 60px;
+  width: 100%;
+  height: 100%;
+  svg {
+    position: absolute;
+    left: 0;
+    top: 0px;
+    height: 45px;
+    width: 100%;
+    margin: $bottomBarTopPadding/2 auto 0 auto;
+  }
+  .iconButtonHighlightSvg{
+    opacity: 0;
+  }
+  &:hover{
+    .iconButtonHighlightSvg{
+      opacity: 1;
+    }
+  }
+  .iconButtonText{
+    position: absolute;
+    top: 45px;
+    left: 0px;
+    width: 100%;
+  }
+
+  &.disabled{
+    opacity: 0.5 !important;
+    filter: grayscale(1);
+  }
+}
