--- src/components/spark-button.scss
@@ -1,49 +1,0 @@
-@import "common.scss";
-
-.sparkButton {
-  min-width: 60px;
-  height: 100%;
-  overflow: hidden;
-  border-bottom-left-radius: 0 !important;
-  border-bottom-right-radius: 0 !important;
-
-  &.disabled{
-    opacity: 0.5 !important;
-  }
-
-
-  .sparkIcon{
-    position: absolute;
-    top: 0;
-    width: 30px;
-    display: block;
-
-    margin: $bottomBarTopPadding auto 0 auto;
-    span{
-      position: absolute;
-      left: 0;
-      height: 38px;
-    }
-
-    .sparkSvg{
-      height: 38px;
-      width: 30px;
-    }
-    .sparkHighlightSvg{
-      height: 38px;
-      width: 30px;
-      opacity: 0;
-    }
-
-  }
-  .sparkText{
-    padding-top: 36px;
-  }
-  &:hover{
-    .sparkIcon{
-      .sparkHighlightSvg{
-        opacity: 1;
-      }
-    }
-  }
-}
