--- src/components/terrain-setup-button.scss
@@ -1,35 +1,0 @@
-@import "common.scss";
-
-.terrainButton {
-  min-width: 60px;
-  height: 100%;
-  overflow: hidden;
-  border-bottom-left-radius: 0 !important;
-  border-bottom-right-radius: 0 !important;
-
-  &.disabled{
-    opacity: 0.5 !important;
-  }
-
-  &:hover svg {
-    box-shadow: 0 0 0 4px rgba(255,255,255,0.5);
-  }
-  &:active svg {
-    box-shadow: 0 0 0 4px rgba(255,255,255,1);
-  }
-
-  &:first-child {
-    border-top-left-radius: 10px;
-  }
-  &:last-child {
-    border-top-right-radius: 10px;
-  }
-  svg {
-    background: white;
-    display: block;
-    height: 32px;
-    width: 60px;
-    border-radius: 0px;
-    margin: 4px auto 0 auto;
-  }
-}
